+++ src/main/java/com/atlauncher/workers/InstanceInstaller.java
@@ -1702,6 +1702,15 @@ public class InstanceInstaller extends SwingWorker<Boolean, Void> implements Net
             libraries.addAll(this.minecraftVersion.libraries);
         }
 
+        // fix Log4J exploits if not on server (servers seem sensitive to this process)
+        if (!isServer) {
+            libraries = libraries.stream().map(library -> {
+                library.fixLog4jVersion();
+
+                return library;
+            }).collect(Collectors.toList());
+        }
+
         return libraries;
     }