+++ src/main/java/com/atlauncher/data/minecraft/loaders/forge/Processor.java
@@ -139,9 +139,9 @@ public class Processor {
                 // checking for local file paths returned "/data/client.lzma" and then makes
                 // sure we localise it to the libraries folder if it's indeed local
                 if (value.charAt(0) == '/') {
-                    if (value.toLowerCase()
-                            .contains(instanceInstaller.root.resolve("libraries").toFile().toString().toLowerCase())
-                            || value.toLowerCase().contains(FileSystem.LIBRARIES.toString().toLowerCase())) {
+                    if (value.toLowerCase().contains(FileSystem.BASE_DIR.toString().toLowerCase())) {
+                        // if the value starts with our base launcher dir, then it's likely already
+                        // resolved (such as {INSTALLER})
                         args.add(value);
                     } else {
                         File localFile = new File(extractedDir, value);