+++ CHANGELOG.md
@@ -3,12 +3,10 @@
 This changelog only contains the changes that are unreleased. For changes for individual releases, please visit the
 [releases](https://github.com/ATLauncher/ATLauncher/releases) page on GitHub.
 
-## 3.4.10.14
+## 3.4.10.15
 
 ### New Features
 
 ### Fixes
-- Processing files from Forge loader on unix systems not working as intended
-- Issue running servers from the launcher that use Forge run.sh files
 
 ### Misc